+++ ecopy/base_funcs/__init__.py
@@ -1,3 +1,4 @@
 from wt_mean import wt_mean
 from wt_var import wt_var
 from wt_scale import wt_scale
+from impute import impute