+++ ecopy/ordination/__init__.py
@@ -5,3 +5,4 @@ from distance import distance
 from transform import transform
 from mds import MDS
 from hillsmith import hillsmith
+from ord_plot import ord_plot