+++ docs/source/changelog.rst
@@ -1,5 +1,17 @@
 Version History
 ==========
+0.0.9
+^^^^^
+- Missing data imputation
+- nls Python 3 compatibility
+- Gower's Euclidean distance for missing data
+- ord_plot function for convex hull and line plots of ordination results
+- Fully incorporated non-linear regression, including documentation
+- Incorporated partial Mantel test in Mantel class
+- Global tests of RDA significance
+- Updated CCA to include correspondence analysis of residual (unconstrained) variance
+- Global tests of CCA significance
+
 0.0.8
 ^^^^^
 - Updated PCA to use SVD instead of eigen decomposition